Add a new, more substancial example written in C.
Add a new example that the fibbonacci sequence, plus some fluff, written in C.
While not perfect, it works.

One issue discovered is the sudden scrambling of elements in arrays if they are not
declared as static; I'm yet to determine if this is a bug in vvm-riscv, GCC or if I've just
messed something up in the Makefile.

Signed-off-by: Vinícius Schütz Piva <vinicius.vsczpv@outlook.com>