Finish IOMAP functionality
While a initial revision, vvm-riscv can now allocate non-uniform memory regions.

Instead of being limited to one huge "ram section", starting at address zero,
now it is possible to allocate arbitrarily sized paged at arbitrary addresses (1KiB aligned).

This did came as a huge overhaul of the memory system, and is very likely to contains bugs.

For example, one can do the following memory mapping:

            allocated      unaloccated      allocated
	[ 0x0000 ~ 2KiB | 0x0800 ~ 60KiB | 0xf800 ~ 512B ]

 * A 2KiB Page is allocated at address zero
 * A 512B Page is allocated at address 0xf800
 * A 60KiB gap (non-allocated) is formed at address 0x800.

Access to non-allocated memory addresses is considered out-of-bounds access.

There is a hardcap on the maximum amount 32 of pages that can be allocated, defined
in the IOMAP_HARDCAP macro, but the number is arbitrary.

Signed-off-by: Vinícius Schütz Piva <vinicius.vsczpv@outlook.com>