Fix off-by-one error in SYS_WRITE stub ecall
The stub ecall would directly add the buffer size to the its address
during the OOB check, meaning that, for example, a one byte message
would be seen as a two byte array.

Specifically, [cpu->regs[11] + size], for example, a buffer @ 0x1000 with 3 bytes,
would calculate it's last address as 0x1000 + 3, which is untrue, as the
last address is actually 0x1002, not 0x1003.

Signed-off-by: Vinícius Schütz Piva <vinicius.vsczpv@outlook.com>